+++ pubky-homeserver/src/data_dir/mod.rs
@@ -4,5 +4,5 @@ mod data_dir;
 mod default_toml;
 mod domain_port;
 
-pub use config_toml::*;
-pub use data_dir::*;
+pub use config_toml::{ConfigToml, ConfigReadError};
+pub use data_dir::DataDir;
