+++ pubky-homeserver/src/lib.rs
@@ -5,11 +5,11 @@
 #![deny(rustdoc::broken_intra_doc_links)]
 #![cfg_attr(any(), deny(clippy::unwrap_used))]
 
-mod config;
+mod config_old;
 mod core;
 mod homeserver;
-mod config2;
+mod data_dir;
 
-pub use config2::*;
+pub use data_dir::*;
 pub use homeserver::Homeserver;
 pub use homeserver::HomeserverBuilder;
